Accept more types in BBoxRotate input_shape argument.

+++ dali/test/python/operator_2/test_bbox_rotate.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# Copyright (c) 2025-2026, NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -316,7 +316,7 @@ def test_allowed_label_shapes(test_labels_shape: list[int]):
             boxes,
             labels,
             angle=45,
-            input_shape=[255, 255],
+            input_shape=np.int32([255, 255]),
             bbox_layout="xyXY",
             bbox_normalized=False,
         )
